Added floor divide and modulus operations, moved tests to tests directory
